% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shr_filter.R
\name{shr_filter}
\alias{shr_filter}
\title{This function filteres data SHR plant growth forms and attached canopy code and \% cover species data from excel workbook sheets.}
\usage{
shr_filter(rawdata)
}
\arguments{
\item{rawdata}{Raw Data.}
}
\value{
Filtered SHR data with attached canopy code and \% cover by species.
}
\description{
This function filteres data SHR plant growth forms and attached canopy code and \% cover species data from excel workbook sheets.
}
\examples{
shr_filter(rawdata)
}
